Run ./gradlew ktlintFormat
