+++ store/src/commonMain/kotlin/org/mobilenativefoundation/store/store5/impl/DefaultLogger.kt
@@ -7,7 +7,6 @@ import org.mobilenativefoundation.store.store5.Logger
  * Default implementation of [Logger] using the Kermit logging library.
  */
 internal class DefaultLogger : Logger {
-
     private val delegate =
         co.touchlab.kermit.Logger.apply {
             setLogWriters(listOf(CommonWriter()))
@@ -18,7 +17,10 @@ internal class DefaultLogger : Logger {
         delegate.d(message)
     }
 
-    override fun error(message: String, throwable: Throwable?) {
+    override fun error(
+        message: String,
+        throwable: Throwable?,
+    ) {
         delegate.e(message, throwable)
     }
 }
